refactored KvHelper into base to make a cleaner model to extend for Certs and Keys
